requirements.txt
Flask==2.2.3
torch==2.1.0
torchvision==0.15.0
Pillow==9.5.0
kagglehub==0.1.1
seaborn==0.12.2
matplotlib==3.7.1
scikit-learn==1.3.0
numpy
pandas